WebGL: Location Area Grid #58

+++ akibot-tanktrack/src/com/akibot/tanktrack/component/world/element/AkiPoint.java
@@ -5,35 +5,35 @@ import java.io.Serializable;
 public class AkiPoint implements Serializable {
 	private static final long serialVersionUID = 1L;
 
-	private float x;
-	private float y;
-	private float z;
+	private double x;
+	private double y;
+	private double z;
 
-	public float getX() {
+	public double getX() {
 		return x;
 	}
 
-	public void setX(float x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 
-	public float getY() {
+	public double getY() {
 		return y;
 	}
 
-	public void setY(float y) {
+	public void setY(double y) {
 		this.y = y;
 	}
 
-	public float getZ() {
+	public double getZ() {
 		return z;
 	}
 
-	public void setZ(float z) {
+	public void setZ(double z) {
 		this.z = z;
 	}
 
-	public AkiPoint(float x, float y, float z) {
+	public AkiPoint(double x, double y, double z) {
 		this.setX(x);
 		this.setY(y);
 		this.setZ(z);
